+++ api/test/unit/services/listing.service.spec.ts
@@ -5,7 +5,6 @@ import {
   LanguagesEnum,
   ListingEventsTypeEnum,
   ListingsStatusEnum,
-  LotteryStatusEnum,
   ReviewOrderTypeEnum,
   UnitTypeEnum,
   UserRoleEnum,
@@ -125,8 +124,6 @@ const requestApprovalMock = jest.fn();
 const changesRequestedMock = jest.fn();
 const listingApprovedMock = jest.fn();
 const lotteryReleasedMock = jest.fn();
-const lotteryPublishedAdminMock = jest.fn();
-const lotteryPublishedApplicantMock = jest.fn();
 
 const canOrThrowMock = jest.fn();
 
@@ -2948,401 +2945,6 @@ describe('Testing listing service', () => {
       expect(prisma.cronJob.update).toHaveBeenCalledWith({
         data: {
           lastRunDate: expect.anything(),
-        },
-        where: {
-          id: expect.anything(),
-        },
-      });
-    });
-  });
-
-  describe('Test expireLotteries endpoint', () => {
-    it('should call the updateMany', async () => {
-      prisma.listings.updateMany = jest.fn().mockResolvedValue({ count: 2 });
-      prisma.cronJob.findFirst = jest
-        .fn()
-        .mockResolvedValue({ id: randomUUID() });
-      prisma.cronJob.update = jest.fn().mockResolvedValue(true);
-
-      await service.expireLotteries();
-      expect(prisma.listings.updateMany).toHaveBeenCalledWith({
-        data: {
-          lotteryStatus: LotteryStatusEnum.expired,
-        },
-        where: {
-          status: ListingsStatusEnum.closed,
-          reviewOrderType: ReviewOrderTypeEnum.lottery,
-          closedAt: {
-            lte: expect.anything(),
-          },
-          OR: [
-            {
-              lotteryStatus: {
-                not: LotteryStatusEnum.expired,
-              },
-            },
-            {
-              lotteryStatus: null,
-            },
-          ],
-        },
-      });
-      expect(prisma.cronJob.findFirst).toHaveBeenCalled();
-      expect(prisma.cronJob.update).toHaveBeenCalled();
-    });
-  });
-
-  describe('Test lotteryStatus endpoint', () => {
-    const adminUser = {
-      id: 'admin id',
-      userRoles: {
-        isAdmin: true,
-      },
-    } as User;
-
-    const partnerUser = {
-      id: 'partner id',
-      userRoles: {
-        isAdmin: false,
-        isPartner: true,
-      },
-    } as User;
-
-    const publicUser = {
-      id: 'partner id',
-      userRoles: {
-        isAdmin: false,
-        isPartner: false,
-      },
-    } as User;
-
-    it('should error when listing is not closed', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.active,
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue(null);
-
-      await expect(
-        async () =>
-          await service.lotteryStatus(
-            {
-              id: randomUUID(),
-              lotteryStatus: LotteryStatusEnum.ran,
-            } as ListingLotteryStatus,
-            user,
-          ),
-      ).rejects.toThrowError(
-        'Lottery status cannot be changed until listing is closed.',
-      );
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        user,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-        },
-      );
-
-      expect(prisma.listings.update).not.toHaveBeenCalled();
-    });
-
-    it.todo(
-      'should not update status if requested status does not match enums',
-    );
-
-    it.todo('should update status to ran from null/errored');
-
-    it('should not update status to ran if user is not an admin', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue(null);
-
-      await expect(
-        async () =>
-          await service.lotteryStatus(
-            {
-              id: randomUUID(),
-              lotteryStatus: LotteryStatusEnum.ran,
-            } as ListingLotteryStatus,
-            partnerUser,
-          ),
-      ).rejects.toThrowError();
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        partnerUser,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-        },
-      );
-
-      expect(prisma.listings.update).not.toHaveBeenCalled();
-    });
-
-    it.todo('should update status to errored');
-
-    it.todo('should update status to approved from ran');
-
-    it.todo('should not update status to approved when status is not ran');
-
-    it.todo('should not update status to approved if user is not an admin');
-
-    it('should update status to releasedToPartners from ran and send email', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.ran,
-        jurisdictionId: 'jurisId',
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.releasedToPartners,
-        jurisdictionId: 'jurisId',
-      });
-
-      jest.spyOn(service, 'getUserEmailInfo').mockResolvedValueOnce({
-        emails: ['admin@email.com', 'partner@email.com'],
-      });
-
-      jest.spyOn(service, 'getPublicUserEmailInfo').mockResolvedValueOnce({
-        en: ['applicant@email.com'],
-      });
-
-      await service.lotteryStatus(
-        {
-          id: randomUUID(),
-          lotteryStatus: LotteryStatusEnum.releasedToPartners,
-        } as ListingLotteryStatus,
-        adminUser,
-      );
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        adminUser,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-          jurisdictionId: 'jurisId',
-        },
-      );
-      expect(prisma.listings.update).toHaveBeenCalledWith({
-        data: {
-          lotteryStatus: LotteryStatusEnum.releasedToPartners,
-        },
-        where: {
-          id: expect.anything(),
-        },
-      });
-
-      expect(service.getUserEmailInfo).toBeCalledWith(
-        ['admin', 'jurisdictionAdmin', 'partner'],
-        'example id',
-        'jurisId',
-      );
-
-      expect(lotteryReleasedMock).toBeCalledWith(
-        { id: 'admin id', userRoles: { isAdmin: true } },
-        { id: 'example id', juris: 'jurisId', name: 'example name' },
-        ['admin@email.com', 'partner@email.com'],
-        config.get('PARTNERS_PORTAL_URL'),
-      );
-    });
-
-    it('should not update status to releasedToPartners if user is not an admin', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.ran,
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue(null);
-      jest.spyOn(service, 'getUserEmailInfo').mockResolvedValueOnce({
-        emails: ['admin@email.com', 'partner@email.com'],
-      });
-      jest.spyOn(service, 'getPublicUserEmailInfo').mockResolvedValueOnce({
-        en: ['applicant@email.com'],
-      });
-
-      await expect(
-        async () =>
-          await service.lotteryStatus(
-            {
-              id: randomUUID(),
-              lotteryStatus: LotteryStatusEnum.releasedToPartners,
-            } as ListingLotteryStatus,
-            partnerUser,
-          ),
-      ).rejects.toThrowError();
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        partnerUser,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-        },
-      );
-
-      expect(prisma.listings.update).not.toHaveBeenCalled();
-    });
-
-    it('should update status to publishedToPublic from releasedToPartners', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.releasedToPartners,
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.publishedToPublic,
-      });
-      jest.spyOn(service, 'getUserEmailInfo').mockResolvedValueOnce({
-        emails: ['admin@email.com', 'partner@email.com'],
-      });
-      jest.spyOn(service, 'getPublicUserEmailInfo').mockResolvedValueOnce({
-        en: ['applicant@email.com'],
-      });
-
-      await service.lotteryStatus(
-        {
-          id: randomUUID(),
-          lotteryStatus: LotteryStatusEnum.publishedToPublic,
-        } as ListingLotteryStatus,
-        partnerUser,
-      );
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        partnerUser,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-        },
-      );
-      expect(prisma.listings.update).toHaveBeenCalledWith({
-        data: {
-          lotteryStatus: LotteryStatusEnum.publishedToPublic,
-        },
-        where: {
-          id: expect.anything(),
-        },
-      });
-    });
-
-    it('should not update status to publishedToPublic when status is not releasedToPartners', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.ran,
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue(null);
-
-      await expect(
-        async () =>
-          await service.lotteryStatus(
-            {
-              id: randomUUID(),
-              lotteryStatus: LotteryStatusEnum.publishedToPublic,
-            } as ListingLotteryStatus,
-            partnerUser,
-          ),
-      ).rejects.toThrowError();
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        partnerUser,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-        },
-      );
-
-      expect(prisma.listings.update).not.toHaveBeenCalled();
-    });
-
-    it('should not update status to publishedToPublic if user is not an admin or partner', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.releasedToPartners,
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue(null);
-
-      await expect(
-        async () =>
-          await service.lotteryStatus(
-            {
-              id: randomUUID(),
-              lotteryStatus: LotteryStatusEnum.publishedToPublic,
-            } as ListingLotteryStatus,
-            publicUser,
-          ),
-      ).rejects.toThrowError();
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        publicUser,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-        },
-      );
-
-      expect(prisma.listings.update).not.toHaveBeenCalled();
-    });
-
-    it('should update status to ran from approved/releasedToPartners/publishedToPublic aka retracted', async () => {
-      prisma.listings.findUnique = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.releasedToPartners,
-      });
-      prisma.listings.update = jest.fn().mockResolvedValue({
-        id: 'example id',
-        name: 'example name',
-        status: ListingsStatusEnum.closed,
-        lotteryStatus: LotteryStatusEnum.ran,
-      });
-      jest.spyOn(service, 'getPublicUserEmailInfo').mockResolvedValueOnce({
-        en: ['applicant@email.com'],
-      });
-
-      await service.lotteryStatus(
-        {
-          id: randomUUID(),
-          lotteryStatus: LotteryStatusEnum.ran,
-        } as ListingLotteryStatus,
-        adminUser,
-      );
-
-      expect(canOrThrowMock).toHaveBeenCalledWith(
-        adminUser,
-        'listing',
-        permissionActions.update,
-        {
-          id: 'example id',
-        },
-      );
-      expect(prisma.listings.update).toHaveBeenCalledWith({
-        data: {
-          lotteryLastRunAt: expect.anything(),
-          lotteryStatus: LotteryStatusEnum.ran,
         },
         where: {
           id: expect.anything(),
