+++ api/test/unit/services/lottery.service.spec.ts
@@ -4,6 +4,7 @@ import {
   ListingsStatusEnum,
   LotteryStatusEnum,
   MultiselectQuestionsApplicationSectionEnum,
+  ReviewOrderTypeEnum,
 } from '@prisma/client';
 import { HttpModule } from '@nestjs/axios';
 import Excel from 'exceljs';
@@ -31,11 +32,24 @@ import MultiselectQuestion from '../../../src/dtos/multiselect-questions/multise
 import { OrderByEnum } from '../../../src/enums/shared/order-by-enum';
 import { LotteryService } from '../../../src/services/lottery.service';
 import { getExportHeaders } from '../../../src/utilities/application-export-helpers';
+import { ListingLotteryStatus } from '../../../src/dtos/listings/listing-lottery-status.dto';
+import { permissionActions } from '../../../src/enums/permissions/permission-actions-enum';
+
+const canOrThrowMock = jest.fn();
+const lotteryReleasedMock = jest.fn();
+const lotteryPublishedAdminMock = jest.fn();
+const lotteryPublishedApplicantMock = jest.fn();
+
+const user = new User();
+user.firstName = 'Test';
+user.lastName = 'User';
+user.email = 'test@example.com';
 
 describe('Testing lottery service', () => {
   let service: LotteryService;
   let prisma: PrismaService;
-  let permissionService: PermissionService;
+  let listingService: ListingService;
+  let config: ConfigService;
 
   beforeAll(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -45,7 +59,12 @@ describe('Testing lottery service', () => {
         MultiselectQuestionService,
         ListingService,
         LotteryService,
-        PermissionService,
+        {
+          provide: PermissionService,
+          useValue: {
+            canOrThrow: canOrThrowMock,
+          },
+        },
         TranslationService,
         ApplicationFlaggedSetService,
         {
@@ -54,8 +73,12 @@ describe('Testing lottery service', () => {
             requestApproval: jest.fn(),
             changesRequested: jest.fn(),
             listingApproved: jest.fn(),
+            lotteryReleased: lotteryReleasedMock,
+            lotteryPublishedAdmin: lotteryPublishedAdminMock,
+            lotteryPublishedApplicant: lotteryPublishedApplicantMock,
           },
         },
+
         ConfigService,
         Logger,
         SchedulerRegistry,
@@ -66,7 +89,8 @@ describe('Testing lottery service', () => {
 
     service = module.get<LotteryService>(LotteryService);
     prisma = module.get<PrismaService>(PrismaService);
-    permissionService = module.get<PermissionService>(PermissionService);
+    listingService = module.get<ListingService>(ListingService);
+    config = module.get<ConfigService>(ConfigService);
   });
 
   describe('Testing lotteryRandomizerHelper()', () => {
@@ -242,7 +266,7 @@ describe('Testing lottery service', () => {
         userRoles: { isAdmin: true },
       } as unknown as User;
 
-      permissionService.canOrThrow = jest.fn().mockResolvedValue(true);
+      canOrThrowMock.mockResolvedValue(true);
       prisma.listings.findUnique = jest.fn().mockResolvedValue({
         id: listingId,
         lotteryLastRunAt: null,
@@ -287,7 +311,7 @@ describe('Testing lottery service', () => {
       await service.lotteryGenerate(
         { user: requestingUser } as unknown as ExpressRequest,
         {} as unknown as Response,
-        { listingId },
+        { id: listingId },
       );
 
       expect(prisma.listings.findUnique).toHaveBeenCalledWith({
@@ -363,7 +387,7 @@ describe('Testing lottery service', () => {
         applicationSet,
         headers,
         {
-          listingId: listingId,
+          id: listingId,
           includeDemographics: false,
           timeZone: 'America/Los_Angeles',
         },
@@ -848,6 +872,413 @@ describe('Testing lottery service', () => {
       expect(workbook.worksheets[0].getRow(3).getCell(3).value).toEqual(4);
       expect(workbook.worksheets[0].getRow(2).getCell(4).value).toEqual('1');
       expect(workbook.worksheets[0].getRow(3).getCell(4).value).toEqual('2');
+    });
+  });
+
+  describe('Testing lotteryStatus()', () => {
+    const adminUser = {
+      id: 'admin id',
+      userRoles: {
+        isAdmin: true,
+      },
+    } as User;
+
+    const partnerUser = {
+      id: 'partner id',
+      userRoles: {
+        isAdmin: false,
+        isPartner: true,
+      },
+    } as User;
+
+    const publicUser = {
+      id: 'partner id',
+      userRoles: {
+        isAdmin: false,
+        isPartner: false,
+      },
+    } as User;
+
+    it('should error when listing is not closed', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.active,
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue(null);
+
+      await expect(
+        async () =>
+          await service.lotteryStatus(
+            {
+              id: randomUUID(),
+              lotteryStatus: LotteryStatusEnum.ran,
+            } as ListingLotteryStatus,
+            user,
+          ),
+      ).rejects.toThrowError(
+        'Lottery status cannot be changed until listing is closed.',
+      );
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        user,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+        },
+      );
+
+      expect(prisma.listings.update).not.toHaveBeenCalled();
+    });
+
+    it.todo(
+      'should not update status if requested status does not match enums',
+    );
+
+    it.todo('should update status to ran from null/errored');
+
+    it('should not update status to ran if user is not an admin', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue(null);
+
+      await expect(
+        async () =>
+          await service.lotteryStatus(
+            {
+              id: randomUUID(),
+              lotteryStatus: LotteryStatusEnum.ran,
+            } as ListingLotteryStatus,
+            partnerUser,
+          ),
+      ).rejects.toThrowError();
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        partnerUser,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+        },
+      );
+
+      expect(prisma.listings.update).not.toHaveBeenCalled();
+    });
+
+    it.todo('should update status to errored');
+
+    it.todo('should update status to approved from ran');
+
+    it.todo('should not update status to approved when status is not ran');
+
+    it.todo('should not update status to approved if user is not an admin');
+
+    it('should update status to releasedToPartners from ran and send email', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.ran,
+        jurisdictionId: 'jurisId',
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.releasedToPartners,
+        jurisdictionId: 'jurisId',
+      });
+
+      jest.spyOn(listingService, 'getUserEmailInfo').mockResolvedValueOnce({
+        emails: ['admin@email.com', 'partner@email.com'],
+      });
+
+      jest
+        .spyOn(listingService, 'getPublicUserEmailInfo')
+        .mockResolvedValueOnce({
+          en: ['applicant@email.com'],
+        });
+
+      await service.lotteryStatus(
+        {
+          id: randomUUID(),
+          lotteryStatus: LotteryStatusEnum.releasedToPartners,
+        } as ListingLotteryStatus,
+        adminUser,
+      );
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        adminUser,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+          jurisdictionId: 'jurisId',
+        },
+      );
+      expect(prisma.listings.update).toHaveBeenCalledWith({
+        data: {
+          lotteryStatus: LotteryStatusEnum.releasedToPartners,
+        },
+        where: {
+          id: expect.anything(),
+        },
+      });
+
+      expect(listingService.getUserEmailInfo).toBeCalledWith(
+        ['admin', 'jurisdictionAdmin', 'partner'],
+        'example id',
+        'jurisId',
+      );
+
+      expect(lotteryReleasedMock).toBeCalledWith(
+        { id: 'admin id', userRoles: { isAdmin: true } },
+        { id: 'example id', juris: 'jurisId', name: 'example name' },
+        ['admin@email.com', 'partner@email.com'],
+        config.get('PARTNERS_PORTAL_URL'),
+      );
+    });
+
+    it('should not update status to releasedToPartners if user is not an admin', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.ran,
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue(null);
+      jest.spyOn(listingService, 'getUserEmailInfo').mockResolvedValueOnce({
+        emails: ['admin@email.com', 'partner@email.com'],
+      });
+      jest
+        .spyOn(listingService, 'getPublicUserEmailInfo')
+        .mockResolvedValueOnce({
+          en: ['applicant@email.com'],
+        });
+
+      await expect(
+        async () =>
+          await service.lotteryStatus(
+            {
+              id: randomUUID(),
+              lotteryStatus: LotteryStatusEnum.releasedToPartners,
+            } as ListingLotteryStatus,
+            partnerUser,
+          ),
+      ).rejects.toThrowError();
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        partnerUser,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+        },
+      );
+
+      expect(prisma.listings.update).not.toHaveBeenCalled();
+    });
+
+    it('should update status to publishedToPublic from releasedToPartners', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.releasedToPartners,
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.publishedToPublic,
+      });
+      jest.spyOn(listingService, 'getUserEmailInfo').mockResolvedValueOnce({
+        emails: ['admin@email.com', 'partner@email.com'],
+      });
+      jest
+        .spyOn(listingService, 'getPublicUserEmailInfo')
+        .mockResolvedValueOnce({
+          en: ['applicant@email.com'],
+        });
+
+      await service.lotteryStatus(
+        {
+          id: randomUUID(),
+          lotteryStatus: LotteryStatusEnum.publishedToPublic,
+        } as ListingLotteryStatus,
+        partnerUser,
+      );
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        partnerUser,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+        },
+      );
+      expect(prisma.listings.update).toHaveBeenCalledWith({
+        data: {
+          lotteryStatus: LotteryStatusEnum.publishedToPublic,
+        },
+        where: {
+          id: expect.anything(),
+        },
+      });
+    });
+
+    it('should not update status to publishedToPublic when status is not releasedToPartners', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.ran,
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue(null);
+
+      await expect(
+        async () =>
+          await service.lotteryStatus(
+            {
+              id: randomUUID(),
+              lotteryStatus: LotteryStatusEnum.publishedToPublic,
+            } as ListingLotteryStatus,
+            partnerUser,
+          ),
+      ).rejects.toThrowError();
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        partnerUser,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+        },
+      );
+
+      expect(prisma.listings.update).not.toHaveBeenCalled();
+    });
+
+    it('should not update status to publishedToPublic if user is not an admin or partner', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.releasedToPartners,
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue(null);
+
+      await expect(
+        async () =>
+          await service.lotteryStatus(
+            {
+              id: randomUUID(),
+              lotteryStatus: LotteryStatusEnum.publishedToPublic,
+            } as ListingLotteryStatus,
+            publicUser,
+          ),
+      ).rejects.toThrowError();
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        publicUser,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+        },
+      );
+
+      expect(prisma.listings.update).not.toHaveBeenCalled();
+    });
+
+    it('should update status to ran from approved/releasedToPartners/publishedToPublic aka retracted', async () => {
+      prisma.listings.findUnique = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.releasedToPartners,
+      });
+      prisma.listings.update = jest.fn().mockResolvedValue({
+        id: 'example id',
+        name: 'example name',
+        status: ListingsStatusEnum.closed,
+        lotteryStatus: LotteryStatusEnum.ran,
+      });
+      jest
+        .spyOn(listingService, 'getPublicUserEmailInfo')
+        .mockResolvedValueOnce({
+          en: ['applicant@email.com'],
+        });
+
+      jest.spyOn(listingService, 'getUserEmailInfo').mockResolvedValueOnce({
+        emails: ['admin@email.com', 'partner@email.com'],
+      });
+
+      await service.lotteryStatus(
+        {
+          id: randomUUID(),
+          lotteryStatus: LotteryStatusEnum.ran,
+        } as ListingLotteryStatus,
+        adminUser,
+      );
+
+      expect(canOrThrowMock).toHaveBeenCalledWith(
+        adminUser,
+        'listing',
+        permissionActions.update,
+        {
+          id: 'example id',
+        },
+      );
+      expect(prisma.listings.update).toHaveBeenCalledWith({
+        data: {
+          lotteryLastRunAt: expect.anything(),
+          lotteryStatus: LotteryStatusEnum.ran,
+        },
+        where: {
+          id: expect.anything(),
+        },
+      });
+    });
+  });
+
+  describe('Test expireLotteries endpoint', () => {
+    it('should call the updateMany', async () => {
+      prisma.listings.updateMany = jest.fn().mockResolvedValue({ count: 2 });
+      prisma.cronJob.findFirst = jest
+        .fn()
+        .mockResolvedValue({ id: randomUUID() });
+      prisma.cronJob.update = jest.fn().mockResolvedValue(true);
+
+      await service.expireLotteries();
+      expect(prisma.listings.updateMany).toHaveBeenCalledWith({
+        data: {
+          lotteryStatus: LotteryStatusEnum.expired,
+        },
+        where: {
+          status: ListingsStatusEnum.closed,
+          reviewOrderType: ReviewOrderTypeEnum.lottery,
+          closedAt: {
+            lte: expect.anything(),
+          },
+          OR: [
+            {
+              lotteryStatus: {
+                not: LotteryStatusEnum.expired,
+              },
+            },
+            {
+              lotteryStatus: null,
+            },
+          ],
+        },
+      });
+      expect(prisma.cronJob.findFirst).toHaveBeenCalled();
+      expect(prisma.cronJob.update).toHaveBeenCalled();
     });
   });
 });
